:recycle: refacto, autocomplete is in another part
