+++ src/pages/AddView.vue
@@ -2,46 +2,25 @@
 import { GeocoderAutocomplete } from '@geoapify/geocoder-autocomplete'
 import { ref, onMounted } from 'vue'
 import { insertSpot } from '../api/spots'
+import SpotsAutocomplete from '../organisms/SpotsAutocomplete.vue'
 
 let newProposal = ref({})
-let suggestions = ref([])
 const friendlyStaff = ref(false)
 const diaperFacilities = ref(false)
 const pmrAccess = ref(false)
 const feedback = ref('')
 const success = ref(false)
-const autoloaderInProgress = ref(false)
-
-const fetchSuggestions = async (e) => {
-  suggestions.value = []
-  if (!autoloaderInProgress.value) {
-    autoloaderInProgress.value = true
-    const response = await fetch(
-      `https://api.geoapify.com/v1/geocode/autocomplete?text=${encodeURIComponent(
-        e.target.value
-      )}&type=amenity&filter=countrycode:fr&limit=5&apiKey=${
-        import.meta.env.VITE_AUTOCOMPLETE_KEY
-      }`,
-      { limit: 5 }
-    )
-    const data = await response.json()
-    const { features } = data
-    for (const item of features) {
-      suggestions.value = [...suggestions.value, item.properties]
-    }
-    autoloaderInProgress.value = false
-  }
-}
+const isSubmissionLoading = ref(false)
 
 const selectSuggestion = (item) => {
   newProposal.value = item
-  suggestions.value = []
 }
 
-const submitForm = (e) => {
+const submitForm = async (e) => {
   event.preventDefault()
   const { value } = newProposal
-  const result = insertSpot({
+  isSubmissionLoading.value = true
+  await insertSpot({
     name: value.name,
     address: value.address_line2,
     lat: value.lat,
@@ -53,6 +32,7 @@ const submitForm = (e) => {
     friendly_staff: friendlyStaff.value
   })
   success.value = true
+  isSubmissionLoading.value = false
   window.scrollTo(0, 0)
   setTimeout(() => (success.value = false), 5000)
 }
@@ -72,27 +52,7 @@ const submitForm = (e) => {
               >Nom et adresse du lieu que vous recommandez ?</label
             >
             <div class="mt-2">
-              <input
-                type="text"
-                name="street-address"
-                id="autocomplete"
-                autocomplete="off"
-                placeholder="Tapez le nom et/ou l'adresse ..."
-                v-model="newProposal.formatted"
-                @input="fetchSuggestions"
-                class="autocomplete-container block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-green-600 sm:text-sm sm:leading-6"
-              />
-              <div v-show="suggestions.length > 0" class="autocomplete-items">
-                <div
-                  :key="index"
-                  v-for="(item, index) in suggestions"
-                  :data-lat="item.lat"
-                  :data-lon="item.lon"
-                  @click="selectSuggestion(item)"
-                >
-                  {{ item.formatted }}
-                </div>
-              </div>
+              <spots-autocomplete :new-proposal="newProposal" @value-selected="selectSuggestion" />
             </div>
           </div>
         </div>
@@ -206,10 +166,36 @@ const submitForm = (e) => {
     <div class="mt-3 flex items-center justify-end gap-x-6">
       <button
         type="submit"
+        v-if="!isSubmissionLoading"
         class="rounded-md bg-green-600 px-3 py-2 text-sm font-semibold text-white shadow-sm hover:bg-green-500 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-green-600"
         @click="submitForm"
       >
         Envoyer
+      </button>
+      <button
+        v-if="isSubmissionLoading"
+        disabled
+        type="button"
+        class="py-2.5 px-5 me-2 text-sm font-medium text-gray-900 bg-white rounded-lg border border-gray-200 hover:bg-gray-100 hover:text-green-700 focus:z-10 focus:ring-2 focus:ring-green-700 focus:text-green-700 dark:bg-gray-800 dark:text-gray-400 dark:border-gray-600 dark:hover:text-white dark:hover:bg-gray-700 inline-flex items-center"
+      >
+        <svg
+          aria-hidden="true"
+          role="status"
+          class="inline w-4 h-4 me-3 text-gray-200 animate-spin dark:text-gray-600"
+          viewBox="0 0 100 101"
+          fill="none"
+          xmlns="http://www.w3.org/2000/svg"
+        >
+          <path
+            d="M100 50.5908C100 78.2051 77.6142 100.591 50 100.591C22.3858 100.591 0 78.2051 0 50.5908C0 22.9766 22.3858 0.59082 50 0.59082C77.6142 0.59082 100 22.9766 100 50.5908ZM9.08144 50.5908C9.08144 73.1895 27.4013 91.5094 50 91.5094C72.5987 91.5094 90.9186 73.1895 90.9186 50.5908C90.9186 27.9921 72.5987 9.67226 50 9.67226C27.4013 9.67226 9.08144 27.9921 9.08144 50.5908Z"
+            fill="currentColor"
+          />
+          <path
+            d="M93.9676 39.0409C96.393 38.4038 97.8624 35.9116 97.0079 33.5539C95.2932 28.8227 92.871 24.3692 89.8167 20.348C85.8452 15.1192 80.8826 10.7238 75.2124 7.41289C69.5422 4.10194 63.2754 1.94025 56.7698 1.05124C51.7666 0.367541 46.6976 0.446843 41.7345 1.27873C39.2613 1.69328 37.813 4.19778 38.4501 6.62326C39.0873 9.04874 41.5694 10.4717 44.0505 10.1071C47.8511 9.54855 51.7191 9.52689 55.5402 10.0491C60.8642 10.7766 65.9928 12.5457 70.6331 15.2552C75.2735 17.9648 79.3347 21.5619 82.5849 25.841C84.9175 28.9121 86.7997 32.2913 88.1811 35.8758C89.083 38.2158 91.5421 39.6781 93.9676 39.0409Z"
+            fill="#1C64F2"
+          />
+        </svg>
+        Loading...
       </button>
     </div>
   </form>
